% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{animate_wordle}
\alias{animate_wordle}
\title{Title}
\usage{
animate_wordle(df, n_tries = 6)
}
\arguments{
\item{df}{tibble with wordle results, must contain x & attempt numeric columns,
and result & letters character columns. For details see example.}

\item{n_tries}{Number of tries to create the empty canvas.}
}
\value{

}
\description{
Title
}
\examples{
first_try_df <-
  tibble(x = rep(1:3),
         letters = c("M", "A", "C"),
         result = c("no", "exact", "almost"),
         attempt = 1)

second_try_df <-
  tibble(x = 1:3,
         letters = c("C", "A", "T"),
         result = rep("exact", 3),
         attempt = 2)

example_df <-
  bind_rows(first_try_df,
            second_try_df)

animate_wordle(example_df)
}
