% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_for_exact}
\alias{check_for_exact}
\title{Check for exact match between a guess and correct answer}
\usage{
check_for_exact(guess, correct)
}
\arguments{
\item{guess}{string Guess for wordle}

\item{correct}{string Correct answer}
}
\value{
The logical vector corresponding to position-wise exact matches between the guess and correct answer. The vector is named with respective letters of a guess, and keeps the order of the characters in the string.
}
\description{
Check for exact match between a guess and correct answer
}
\examples{
check_for_exact("stop", "start")
check_for_exact("aloft", "float")
check_for_exact("test", "test")
}
