% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_for_app.R
\name{check_guess}
\alias{check_guess}
\title{Check guess}
\usage{
check_guess(word, words_df = words_for_wordle_df)
}
\arguments{
\item{word}{string}

\item{words_df}{tibble with words for wordle, must have word character column}
}
\value{
logical if word is among the words
}
\description{
Check guess
}
\examples{
check_guess("start") # TRUE
check_guess("yolo") # FALSE

}
