% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessing.R
\name{guess_first}
\alias{guess_first}
\title{Select first word}
\usage{
guess_first(words_df)
}
\arguments{
\item{words_df}{tibble with preselected (for example based on length) words}
}
\value{
String with a guess word
}
\description{
Select first word
}
\examples{
data(words_for_wordle)
guess_first(words_for_wordle)
}
