% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessing.R
\name{guess_word}
\alias{guess_word}
\title{Guess word and narrow down the guessing space}
\usage{
guess_word(words_df, previous_guess, regexpr_word, regexpr_letters)
}
\arguments{
\item{words_df}{tibble with pre-selected (for example based on length) words}

\item{previous_guess}{string with last guess}

\item{regexpr_word}{string with regular expression combining knowledge
from previous guesses has}

\item{regexpr_letters}{string with regular expression with screened letters;
for example to screen for "a" and "c" the string would look like this "a|c"}
}
\value{
named list with:
filtered tibble with possible guesses left and
string with current guess.
}
\description{
Guess word and narrow down the guessing space
}
\examples{

words_df <- words_for_wordle_df \%>\% filter(word \%in\% c("cat", "man", "can"))
previous_guess <- "can"
regexpr_word <- "[c]{1}[a]{1}[abcdefghijklmopqrstuvwxyz]{1}"
regexpr_letters <- ""
guess_word(words_df, previous_guess, regexpr_word, regexpr_letters)
}
