% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_from_guess.R
\name{learn_from_guess}
\alias{learn_from_guess}
\title{Learn from the guess}
\usage{
learn_from_guess(testing_vec, allowed, scan_for)
}
\arguments{
\item{testing_vec}{named character vector; output of \code{test_guess}
function.}

\item{allowed}{list of lists with all previously learned information about
possible letters at given position. Starting with all letters available
at all positions. Information learned from new guess saved as new element.
For more details see Examples.}

\item{scan_for}{character vector, subset of \code{letters}.}
}
\value{
Named list with updated \code{allowed} and \code{scan_for} objects.
}
\description{
Learn from the guess
}
\examples{
test_vec <- test_guess("aloft", "float")
# this is first guess, hence no prior knowledge
allowed_letters <- list(lapply(1:5, function(i) letters))
scan_for <- c()

learn_from_guess(test_vec, allowed_letters, scan_for)
}
