% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_for_app.R
\name{plot_keybord}
\alias{plot_keybord}
\title{Plot keyboard outline with selected "grayed" out letters}
\usage{
plot_keybord(used_letters)
}
\arguments{
\item{used_letters}{character vector with letters to grey out}
}
\value{
ggplot with keyboard outline
}
\description{
Plot keyboard outline with selected "grayed" out letters
}
\examples{
plot_keybord(c()) # all green
plot_keybord(c("a", "b", "t")) # a, b and t grayed out
}
