% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_wordle}
\alias{plot_wordle}
\title{Plot wordle guesses}
\usage{
plot_wordle(df = NULL, n_tries = 6, n_letters = NULL)
}
\arguments{
\item{df}{tibble with wordle results, must contain x & y numeric columns,
and result, letters and attempt character columns. For details see example.}

\item{n_tries}{numeric, number of allowed guesses}

\item{n_letters}{numeric, number of letters}
}
\value{
ggplot2 plot with visualized guesses
}
\description{
Plot wordle guesses
}
\examples{

first_try_df <-
  tibble(x = rep(1:3),
         letters = c("M", "A", "C"),
         result = c("no", "exact", "almost"),
         attempt = 1)

second_try_df <-
  tibble(x = 1:3,
         letters = c("C", "A", "T"),
         result = rep("exact", 3),
         attempt = 2)

example_df <-
  bind_rows(first_try_df,
            second_try_df)

plot_wordle(first_try_df)
plot_wordle(example_df)
}
