% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_for_app.R
\name{praise_success}
\alias{praise_success}
\title{Praise successful solution}
\usage{
praise_success(i)
}
\arguments{
\item{i}{integer, successful try}
}
\value{
string praise for successfully solving wordle
}
\description{
Praise successful solution
}
\examples{
praise_success(2)
}
