% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_for_app.R
\name{run_wordle_guess}
\alias{run_wordle_guess}
\title{Run wordle guess, check it, store results and plot the viz}
\usage{
run_wordle_guess(
  testing_vec = NULL,
  i = 1,
  guesses_df = tibble(),
  n_letters = NULL,
  n_tries = NULL
)
}
\arguments{
\item{testing_vec}{named character vector, output of \code{test_guess}
function}

\item{i}{numeric, guess number}

\item{guesses_df}{tibble, saved previous guesses}

\item{n_letters}{numeric, length of guess and answer}

\item{n_tries}{numeric, maximum number of tries}
}
\value{
list with updated guesses_df (if \code{testing_vec} provided),
animation (if \code{gganimate} installed) and a static plot
}
\description{
Run wordle guess, check it, store results and plot the viz
}
\examples{
run_wordle_guess(n_letters = 5, n_tries = 6)
}
