% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate one game}
\usage{
simulate(words_for_wordle_df, n_letters = 5, n_guesses = 6)
}
\arguments{
\item{words_for_wordle_df}{tibble with words}

\item{n_letters}{integer with the choice of the size of words, default: 5}

\item{n_guesses}{integer with the number of guesses allowed, default: 6}
}
\value{
tibble with results of the simulation
}
\description{
Simulate one game
}
\examples{
simulate(words_for_wordle_df, 3)
}
