% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{test_guess}
\alias{test_guess}
\title{Test a guess against correct answer}
\usage{
test_guess(guess, correct)
}
\arguments{
\item{guess}{A string.}

\item{correct}{A string.}
}
\value{
The character vector corresponding to position-wise feedback. "no" means the letter is not found in the word, "almost" suggest a different position and "exact" confirms that a letter is in it's right position. The vector is named with respective letters of a guess, and keeps the order of the characters in the string.
}
\description{
Test a guess against correct answer
}
\examples{
test_guess("stops", "start")
test_guess("aloft", "float")
test_guess("test", "test")
}
